+++ bs_cli/adbs_cli/entry_point_commands.py
@@ -458,13 +458,13 @@ def test(component: str, branch: str, quick: bool, main: bool, verbose: bool=Tru
 @click.option("-i", "--ioc", required=False, help="Deploy only to the specified ioc(s). If 'ALL', all iocs in facilities specified by facility arg will be deployed. Seperate iocs by comma, ex: sioc-sys0-test1,sioc-sys0-test2.")
 @click.option("-t", "--tag", required=False, help="Component tag to deploy")
 @click.option("-ls", "--list", is_flag=True, required=False, help="List the active releases")
-@click.option("-l", "--local", is_flag=True, required=False, help="Deploy local directory instead of the artifact storage")
+# @click.option("-l", "--local", is_flag=True, required=False, help="Deploy local directory instead of the artifact storage")
 @click.option("-r", "--revert", is_flag=True, required=False, help="Revert to previous version")
 @click.option("-u", "--update-db", is_flag=True, required=False, help="Add a new deployment to the deployment configuration/database (Only use when deploying app/iocs for the first time)")
 # @click.option("-o", "--override", is_flag=True, required=False, help="Point local DEV deployment to your user-space repo")
 @click.option("-n", "--dry-run", is_flag=True, required=False, help="Print the commands that would be executed, but do not execute them.")
 @click.option("-v", "--verbose", is_flag=True, required=False, help="More detailed output")
-def deploy(component: str, facility: str, test: bool, ioc: str, tag: str, list: bool, local: bool,
+def deploy(component: str, facility: str, test: bool, ioc: str, tag: str, list: bool,
             revert: bool, update_db: bool, dry_run: bool, verbose: bool):
     """Trigger a deployment. Automatically deploys app and ioc(s) to the tag you choose. Facility is automatically determined by ioc.
         Will automatically pickup app in the directory you're sitting in.
@@ -484,8 +484,8 @@ def deploy(component: str, facility: str, test: bool, ioc: str, tag: str, list: 
         under_development()
 
     # 1.2) Option - local
-    if (local):
-        under_development()
+    # if (local):
+    #     under_development()
 
     # 1.3) Option - revert
     if (revert):